+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Slightly increased the padding to the right of the plot when `scale_ggsurvfit()` is called. (#165)
 
+* Now exporting the utility function `align_plots()` that is used to ensure risktable and the primary plot align. By exporting this function, users will now be able to construct custom risktable graphics. See `?align_plots()` for an example. (#175)
+
 # ggsurvfit 0.3.1
 
 * For transformations in `tidy_survfit()` that change the monotonicity of the curve, the `conf.low` and `conf.high` column names are now switched. (#154)